refactor: always render with a view (#104)
Implement an abstract interface for editor views and make a new concrete
implementation of a static editor view to use in server-side rendering
and before the document mounts.

Refactor the editor view to be a proper sub-class of its base class:

- Declare private members from the base class.

- Remove the React-managed DOM when calling the base class constructor,
  and restore it afterwards.

- Track the next props, rather than the last props, overriding all of
  the exposed methods for updating props and state to be pure. Use them
  to maintain an up-to-date node view set so that the base class never
  redraws the document view.

- Delegate to the base class update method to commit new props. Remove
  the explicit calls to set the selection and let the base handle it.
  Remove the explicit remounting of the view when plugin views change,
  because the base class now handles updating plugin views.

- Remove the SSR stubs, which should no longer be needed.

Throw an invalid state error when attempting to invoke an editor event
callback without a mounted document.