Prevent double-enter handlers on iOS (#119)
#72 added a beforeInput handler for insertParagraph/insertLineBreak input types. On many OS/browser combinations, the keydown event for Enter would have been triggered and stopped the event before the beforeInput handler runs, but on Chrome on Android, all inputs are part of an IME composition, and so no Enter event code is ever emitted (the key code is always 229, instead).

It turns out that prosemirror-view has a hack that runs the Enter event handlers on iOS, to avoid a bug in Safari with calling event.preventDefault() on Enter events. This results in Enter handlers executing twice, first in our beforeInput handler, and then again in the deferred prosemirror-view handler.

Luckily, there is a flag that prosemirror-view sets and then checks in the deferred handler. So now we unset that flag in our beforeInput handler, which runs before the deferred handler checks the flag, preventing the deferred handler from running.