Move `view.destroy()` call into layout effect body.
In React 19, in StrictMode, in some circumstances, React will clean-up and
re-run layout effects even if their dependencies have not changed and the
component has not remounted.

This makes it unsafe to run `view.destroy()` in the clean-up for a layout
effect that does not re-create the view, because the view can be destroyed
even when it is not about to be updated to a new value, leaving a destroyed
view in the document.

To resolve this, we move `view.destroy()` into the body of the layout effect
that is actually responsible for creating the new view, and only when
we're certain that we're about to create a new view.