Add a destroyed property to ReactEditorView
When a render occurs that would destroy and recreate the
EditorView, we still run a layout effect that attempts to
commit pending effects for the now-destroyed layout effect.
ProseMirror View, in this scenario, just re-attaches the
DOM Observer and its selectionchange listener (but not the
rest of the input event listeners). This results in
two separae selectionchange handlers, one for a stale EditorView.

To prevent this, we track whether an EditorView has been
destroyed, and abort from commitPendingEffects if it has.